% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean.data}
\alias{clean.data}
\title{cleans sensor data with user-specified routines}
\usage{
clean.data(deploy = "pellerin", folder = "../examples/",
  plot.diagnostic = TRUE, write.file = FALSE)
}
\arguments{
\item{deploy}{a string for the *.yml file name}

\item{folder}{a string which specifies the folder of the *.yml file}

\item{plot.diagnostic}{a boolean for creating a diagnostic plot}

\item{write.file}{a boolean for creating a timeseries file output}
}
\value{
An optional plot and optional file output handle
}
\description{
cleans sensor data according to details within the user-specified config (*yml) file  \cr
}
\examples{
\dontrun{
clean.data(deploy='pellerin',folder='/[User]/Desktop/Science Projects/sensorQC/',plot.diagnostic=FALSE)
clean.data(deploy='pellerin',folder='/[User]/Desktop/Science Projects/sensorQC/',plot.diagnostic=TRUE,write.file=TRUE)
}
}
\author{
Jordan S. Read
}
\keyword{methods}

