% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{clean a sensor or a flagged sensor}
\usage{
clean(x, which, ..., replace)
}
\arguments{
\item{x}{sensor object}

\item{which}{flags to replace}

\item{replace}{replacement value for flagged indices. 
NULL removes the flagged row}

\item{\dots}{additional flags to use (passed to flag(x,...))}
}
\description{
clean a sensor or a flagged sensor
}
\examples{
data = c(999999, 1,2,3,4,2,3,4)
sensor = flag(data, 'x > 9999')
clean(sensor)
clean(data, 'x > 9999', 'persist(x) > 10', 'MAD(x) > 3', replace=NA)
}

