% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_data}
\alias{clean_data}
\title{cleans sensor data with user-specified routines}
\usage{
clean_data(deploy, folder, plot_diagnostic = TRUE, write_file = FALSE)
}
\arguments{
\item{deploy}{a string for the *.yml file name}

\item{folder}{a string which specifies the folder of the *.yml file}

\item{plot_diagnostic}{a boolean for creating a diagnostic plot}

\item{write_file}{a boolean for creating a timeseries file output}
}
\value{
An optional plot and optional file output handle
}
\description{
cleans sensor data according to details within the user-specified config (*yml) file  \cr
}
\examples{
folder <- system.file('extdata', package = 'sensorQC')
clean_data(deploy = 'pellerin', folder = folder, plot_diagnostic=FALSE)
clean_data(deploy = 'pellerin', folder = folder, plot_diagnostic=TRUE, write_file=TRUE)
clean_data(deploy = 'no_window', folder = folder, plot_diagnostic=FALSE, write_file=FALSE)
}
\author{
Jordan S. Read
}
\keyword{methods}

