% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compress_flags}
\alias{compress_flags}
\title{compresses boolean flags into int matrix padded with NAs
flags are compressed because it is assumed that they don't happen incredibly frequently
example: flag.bool <- matrix(nrow=3,ncol=4,data=c(F,F,F,F,F,F,T,T,F,F,F,F))}
\usage{
compress_flags(flag.bool)
}
\description{
compresses boolean flags into int matrix padded with NAs
flags are compressed because it is assumed that they don't happen incredibly frequently
example: flag.bool <- matrix(nrow=3,ncol=4,data=c(F,F,F,F,F,F,T,T,F,F,F,F))
}

