% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_sensor}
\alias{load_sensor}
\title{Loads sensor data into data.frame}
\usage{
load_sensor(file, format = "wide_burst", date_type)
}
\arguments{
\item{file}{a string with path to the sensor data}

\item{format}{a string which matches a valid sensor data format}

\item{date_type}{a string which represents a valid date format type}
}
\value{
A data.frame with DateTime and values
}
\description{
loads sensor data into data.frame according to the file format specified.  \cr
}
\examples{
file <- system.file('extdata', 'test_data.txt', package = 'sensorQC')
data_out <- load_sensor(file, format="wide_burst", date_type="\%m/\%d/\%Y \%H:\%M")
print(data_out)
}
\author{
Jordan S. Read
}
\keyword{methods}

