% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_sqc.R
\name{load_sqc}
\alias{load_sqc}
\title{load in configuration file for sensorQC}
\usage{
load_sqc(deploy.name, folder)
}
\arguments{
\item{deploy.name}{name of config file (proceeding *.yml extension).}

\item{folder}{folder path for config file}
}
\value{
a sqc object
}
\description{
Loads text file for sensorQC. Assumes *.yml file extension for deployment.  \cr
}
\details{
a \code{sensorQC} function loading config.\cr
}
\examples{
folder <- system.file('extdata', package = 'sensorQC')
sqc <- load_sqc(deploy.name='pellerin',folder = folder)
}
\author{
Jordan S. Read
}
\keyword{methods}

