% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/custom-functions.R
\name{MAD}
\alias{MAD}
\title{median absolute deviation outlier test}
\usage{
MAD(x, w)
}
\arguments{
\item{x}{values}

\item{windows}{vector of equal length to x specifying windows}
}
\value{
a vector of MAD normalized values relative to an undefined rejection criteria (usually 2.5 or 3).
}
\description{
median absolute deviation outlier test
}
\author{
Jordan S. Read
}
\keyword{MAD}

