\name{median.absolute.deviation}
\alias{median.absolute.deviation}
\title{median absolute deviation outlier test}
\usage{
  median.absolute.deviation(data.in)
}
\arguments{
  \item{\code{data.in}}{a \code{sensorQC} data.frame.}
}
\value{
  a vector of MAD normalized values relative to an
  undefined rejection criteria (usually 2.5 or 3).
}
\description{
  median absolute deviation outlier test
}
\author{
  Jordan S. Read
}
\keyword{MAD}

