% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-functions.R
\name{persist}
\alias{persist}
\title{function for checking persistent values}
\usage{
persist(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a vector that provides a count of persistance
}
\description{
repeated values of a vector
}

