% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{read.default}
\title{read in a sensor or configuration file}
\usage{
read(file, ...)

read.default(file, format, date.format, ...)
}
\description{
read from file and create sensor or qconfig object
}
\examples{
qconfig = read(system.file('extdata','pedro.yml', package = 'sensorQC'))
file <- system.file('extdata', 'test_data.txt', package = 'sensorQC')
data_out <- read(file, format="wide_burst", date.format="\%m/\%d/\%Y \%H:\%M")
}

