% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sensor-class.R
\name{sensor}
\alias{sensor}
\title{create a sensor object}
\usage{
sensor(x, flag.defs, ...)
}
\arguments{
\item{x}{a data.frame or sensor object}

\item{flag.defs}{definitions for flags. Functions that can be evaluated to T/F}

\item{\dots}{additional flag.defs}
}
\description{
sensor data
}

