% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window.sensor}
\alias{window}
\alias{window.sensor}
\title{window sensorQC data}
\usage{
\method{window}{sensor}(x, ..., type)
}
\arguments{
\item{x}{a data.frame of time series data}

\item{type}{numeric, in seconds, specifying the window time width. or 
"auto" to automatically window data}

\item{\dots}{not used}
}
\value{
a list of time series data and indices for breaks
}
\description{
Breaks up time series data into window chunks.  \cr
}
\author{
Jordan S. Read
}
\keyword{methods}

