% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/window_data.R
\name{window_data}
\alias{window.sensor}
\alias{window_data}
\title{window sensorQC data}
\usage{
\method{window}{sensor}(x, window, ...)
}
\arguments{
\item{window}{numeric, in seconds, specifying the window time width. or
"auto" to automatically window data}

\item{data.in}{a data.frame of time series data}
}
\value{
a list of time series data and indices for breaks
}
\description{
Breaks up time series data into window chunks.  \cr
}
\author{
Jordan S. Read
}
\keyword{window}

