% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{Creates flag vector based on input data}
\usage{
flag(x, flag.defs, ...)
}
\arguments{
\item{x}{data}

\item{flag.defs}{definitions for flagging}

\item{\dots}{additional defs for flags}
}
\value{
a vector of flags of length equal to number of rows in data.in
}
\description{
Creates flag vector with codes and methods according to params list.  \cr
}
\examples{
dates <- seq(as.POSIXct('1999-01-01'),by=1,length.out=14)
values <- c(runif(12,2,4),NA,NA)
data.in <- data.frame("DateTime"=dates,"sensor.obs"=values)
simple.sqc <- list(list(expression="x == 999999",type="error_code",description="logger error code"),
             list(expression='is.na(x)',type='error_code',description='missing data'))

flag(data.in, "x == 999999")
flag(data.in,simple.sqc)
}
\author{
Jordan S. Read
}
\keyword{methods}

