% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.sensor}
\alias{load.sensor}
\title{Loads sensor data into data.frame}
\usage{
load.sensor(filename = "../examples/test_data.txt", format = "Pellerin")
}
\arguments{
\item{filename}{a string}

\item{format}{a string which matches a valid sensor data format}
}
\value{
A data.frame with DateTime and values
}
\description{
loads sensor data into data.frame according to the file format specified.  \cr
}
\examples{
filename <- '../examples/test_data.txt'
data.out <- load.sensor(filename,format="Pellerin")
print(data.out)
}
\author{
Jordan S. Read
}
\keyword{sqc}

