% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MAD}
\alias{MAD}
\title{median absolute deviation outlier test}
\usage{
MAD(data.in)
}
\arguments{
\item{data.in}{a \code{sensorQC} data.frame.}
}
\value{
a vector of MAD normalized values relative to an undefined rejection criteria (usually 2.5 or 3).
}
\description{
median absolute deviation outlier test
}
\author{
Jordan S. Read
}
\keyword{MAD}

