% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flag_functions.R
\name{MAD}
\alias{MAD}
\title{median absolute deviation outlier test}
\usage{
MAD(vals, windows)
}
\arguments{
\item{data.in}{a \code{sensorQC} data.frame.}
}
\value{
a vector of MAD normalized values relative to an undefined rejection criteria (usually 2.5 or 3).
}
\description{
median absolute deviation outlier test
}
\author{
Jordan S. Read
}
\keyword{MAD}

