% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\alias{read.default}
\title{read in a sensor or configuration file}
\usage{
read(file, ...)

read.default(file, format, date.format, ...)
}
\arguments{
\item{file}{filename}

\item{format}{format for file}

\item{date.format}{data format of file}

\item{\dots}{additional arguments passed to file read methods}
}
\description{
read from file and create sensor or qconfig object

read in sensor data
}
\examples{
qconfig = read(system.file('extdata','pedro.yml', package = 'sensorQC'))
}

